조회/장바구니추가 및 삭제 데이터 삽입기능 추가

+++ mainProject/web/main/src/main/java/kr/co/test/vo/CartHistoryVO.java
@@ -5,8 +5,8 @@ import lombok.Data;
 @Data
 public class CartHistoryVO {
 	private Long cartHistoryId;
-    private Long userId;
-    private Long productId;
+    private int userId;
+    private int productId;
     private String action; // 'ADD' or 'REMOVE'
     private java.util.Date actionDate;
 }
