+++ mainProject/web/main/src/main/java/kr/co/test/vo/UserViewVO.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class UserViewVO {
-	private Long viewId;
-    private Long userId;
-    private Long productId;
+	private int viewId;
+    private int userId;
+    private int productId;
     private java.util.Date viewedAt;
 }
